+++ syft/workers/__init__.py
@@ -6,5 +6,7 @@
 from .abstract import AbstractWorker  # noqa: F401
 from .base import BaseWorker  # noqa: F401
 from .virtual import VirtualWorker  # noqa: F401
+from .websocket_worker import WebsocketClientWorker  # noqa: F401
+from .websocket_worker import WebsocketServerWorker  # noqa: F401
 
-__all__ = ["base", "virtual"]
+__all__ = ["base", "virtual", "websocket_worker"]
